addressed comments from layershifter
