--- docs/src/examples/components/Dropdown/Types/DropdownExampleMultipleSearch.shorthand.tsx
@@ -1,31 +1,0 @@
-import * as React from 'react'
-import { Dropdown } from '@stardust-ui/react'
-
-const inputItems = [
-  'Bruce Wayne',
-  'Natasha Romanoff',
-  'Steven Strange',
-  'Alfred Pennyworth',
-  `Scarlett O'Hara`,
-  'Imperator Furiosa',
-  'Bruce Banner',
-  'Peter Parker',
-  'Selina Kyle',
-]
-
-const DropdownExampleMultipleSearch = () => (
-  <Dropdown
-    search
-    multiple
-    items={inputItems}
-    placeholder="Start typing a name"
-    getA11ySelectionMessage={getA11ySelectionMessage}
-    noResultsMessage="We couldn't find any matches."
-  />
-)
-
-const getA11ySelectionMessage = {
-  onAdd: item => `${item} has been selected.`,
-  onRemove: item => `${item} has been removed.`,
-}
-export default DropdownExampleMultipleSearch
